Replace IPC with Redux Saga channel
